bringup: Step 21 code modularisation — split overgrown files into per-subsystem modules
5-phase systematic split aligned with planned Core 1 HAL architecture:
- Phase 1: bringup.h (340→127) → 6 sub-headers (pins, lm27965, bq25622, bq27441, sx1262, power)
- Phase 2: bringup_power.c (1076→168) → led.c, charger.c, gauge.c
- Phase 3: bringup_sram.c (1721→86) → psram.c, memory_tft.c, flash.c expanded
- Phase 4: bringup_audio.c (765→deleted) → amp.c, mic.c
- Phase 5: bringup_menu.c (778→414) → tft_draw.c
Bringup log reorganised: removed stale "Pending Steps" section, TFT layouts
moved to tft-layouts.md. Zero behaviour change; build+flash verified.
