+++ firmware/tools/bringup/bringup_sram.c
@@ -84,1638 +84,3 @@ void sram_test(void) {
     printf("  Result: %s\n", total_errors == 0 ? "PASS" : "FAIL");
 }
 
-// ---------------------------------------------------------------------------
-// PSRAM test (APS6404L, 8 MB, CS = GPIO 0 / QMI CS1n)
-//
-// pico-sdk 2.2.0 has no hardware/psram.h — use QMI registers directly.
-// Physical address map (RP2350B, 16 MB flash on CS0):
-//   CS0 physical 0x000000..0xFFFFFF  (flash)
-//   CS1 physical 0x800000..0xFFFFFF  (PSRAM, 8 MB on M1)
-// Default ATRANS[2,3] already map virtual 0x800000..0xFFFFFF → M1.
-// XIP uncached access: XIP_NOCACHE_NOALLOC_BASE + 0x800000.
-// ---------------------------------------------------------------------------
-
-#define PSRAM_TEST_WORDS 1024  // 4 KB
-
-// psram_test — XIP read/write test. Requires psram_init() to have been called.
-void psram_test(void) {
-    printf("\n--- PSRAM XIP test (APS6404L, CS=GPIO%d) ---\n", PSRAM_CS_PIN);
-
-    // Verify GPIO0 is XIP_CS1
-    uint32_t gpio0_ctrl = *(volatile uint32_t *)0x40028004u;
-    uint32_t funcsel = gpio0_ctrl & 0x1Fu;
-    printf("  GPIO0 FUNCSEL=%u %s\n", (unsigned)funcsel,
-           funcsel == 9 ? "(XIP_CS1 OK)" : "(WRONG — run psram_init first)");
-    if (funcsel != 9) return;
-
-    printf("  M1: timing=0x%08X rfmt=0x%08X rcmd=0x%02X wfmt=0x%08X wcmd=0x%02X\n",
-           (unsigned)qmi_hw->m[1].timing, (unsigned)qmi_hw->m[1].rfmt,
-           (unsigned)qmi_hw->m[1].rcmd,
-           (unsigned)qmi_hw->m[1].wfmt, (unsigned)qmi_hw->m[1].wcmd);
-    printf("  XIP_CTRL WRITABLE_M1=%u\n",
-           (xip_ctrl_hw->ctrl & XIP_CTRL_WRITABLE_M1_BITS) ? 1 : 0);
-
-    volatile uint32_t *psram = (volatile uint32_t *)PSRAM_NOCACHE;
-
-    // SPI write+readback (two words)
-    psram[0] = 0xA55A1234u;
-    psram[1] = 0x12345678u;
-    uint32_t rd0 = psram[0];
-    uint32_t rd1 = psram[1];
-    bool pair_ok = (rd0 == 0xA55A1234u) && (rd1 == 0x12345678u);
-    printf("  XIP wr+rd [0]: wr=0xA55A1234 rd=0x%08X\n", (unsigned)rd0);
-    printf("  XIP wr+rd [1]: wr=0x12345678 rd=0x%08X\n", (unsigned)rd1);
-    printf("  XIP pair: %s\n", pair_ok ? "PASS" : "FAIL");
-
-    // Pattern test (4 KB)
-    for (int i = 0; i < PSRAM_TEST_WORDS; i++)
-        psram[i] = (uint32_t)(0xA5000000u | (uint32_t)i);
-
-    int errors = 0;
-    for (int i = 0; i < PSRAM_TEST_WORDS; i++) {
-        uint32_t expected = (uint32_t)(0xA5000000u | (uint32_t)i);
-        if (psram[i] != expected) {
-            if (errors < 4)
-                printf("  ERR @ +0x%04X: wr 0x%08X rd 0x%08X\n",
-                       i * 4, (unsigned)expected, (unsigned)psram[i]);
-            errors++;
-        }
-    }
-    printf("  Pattern (4 KB): %s", errors == 0 ? "PASS" : "FAIL");
-    if (errors) printf(" (%d errors)", errors);
-    printf("\n");
-}
-
-// ---------------------------------------------------------------------------
-// PSRAM CS diagnostic — Issue 8: GPIO0 stuck LOW investigation
-//
-// Tests:
-//   1. Read GPIO0 pad input level and FUNCSEL before any changes
-//   2. Read QMI direct_csr ASSERT_CS1N bit
-//   3. Override GPIO0 to SIO output HIGH — check if MCU can drive it
-//   4. Drive LOW, then HIGH again — confirm MCU control
-//   5. Restore to XIP_CS1 function
-// ---------------------------------------------------------------------------
-void psram_diag(void) {
-    printf("\n--- PSRAM CS diagnostic (GPIO0, Issue 8) ---\n");
-
-    // 1. Read current GPIO0 state before touching anything
-    uint32_t gpio0_ctrl = *(volatile uint32_t *)0x40028004u;  // IO_BANK0 GPIO0_CTRL
-    uint32_t gpio0_status = *(volatile uint32_t *)0x40028000u;  // IO_BANK0 GPIO0_STATUS
-    uint32_t pads0 = *(volatile uint32_t *)0x40038004u;  // PADS_BANK0 GPIO0
-    uint32_t funcsel = gpio0_ctrl & 0x1Fu;
-    bool pad_in = (gpio0_status >> 17) & 1u;  // INFROMPAD (bit 17)
-    bool pad_out = (gpio0_status >> 9) & 1u;  // OUTTOPAD (bit 9)
-    bool pad_oe = (gpio0_status >> 13) & 1u;  // OETOPAD (bit 13)
-
-    printf("  [Before] GPIO0_CTRL=0x%08X  FUNCSEL=%u (%s)\n",
-           (unsigned)gpio0_ctrl, (unsigned)funcsel,
-           funcsel == 9 ? "XIP_CS1" :
-           funcsel == 5 ? "SIO" :
-           funcsel == 31 ? "NULL" : "other");
-    printf("  [Before] GPIO0_STATUS=0x%08X\n", (unsigned)gpio0_status);
-    printf("           INFROMPAD=%u  OUTTOPAD=%u  OETOPAD=%u\n",
-           pad_in, pad_out, pad_oe);
-    printf("  [Before] PADS_BANK0_GPIO0=0x%08X (OD=%u PUE=%u PDE=%u IE=%u)\n",
-           (unsigned)pads0,
-           (pads0 >> 7) & 1, (pads0 >> 3) & 1,
-           (pads0 >> 2) & 1, (pads0 >> 6) & 1);
-
-    // 2. Read QMI direct_csr — check if CS1n is asserted
-    uint32_t direct_csr = qmi_hw->direct_csr;
-    printf("  [Before] QMI direct_csr=0x%08X  EN=%u  ASSERT_CS1N=%u\n",
-           (unsigned)direct_csr,
-           (direct_csr >> 0) & 1u,
-           (direct_csr >> 2) & 1u);
-
-    // 3. Override to SIO output HIGH — does the pin actually go HIGH?
-    gpio_init(PSRAM_CS_PIN);
-    gpio_set_dir(PSRAM_CS_PIN, GPIO_OUT);
-    gpio_put(PSRAM_CS_PIN, 1);
-    busy_wait_ms(1);
-
-    gpio0_status = *(volatile uint32_t *)0x40028000u;
-    pad_in = (gpio0_status >> 17) & 1u;
-    pad_out = (gpio0_status >> 9) & 1u;
-    printf("  [SIO HIGH] OUTTOPAD=%u  INFROMPAD=%u  --> %s\n",
-           pad_out, pad_in,
-           pad_in ? "MCU can drive HIGH (no external short)" :
-                    "STILL LOW — external short to GND!");
-
-    // 4. Drive LOW, confirm it goes LOW
-    gpio_put(PSRAM_CS_PIN, 0);
-    busy_wait_ms(1);
-    gpio0_status = *(volatile uint32_t *)0x40028000u;
-    pad_in = (gpio0_status >> 17) & 1u;
-    printf("  [SIO LOW]  INFROMPAD=%u  --> %s\n",
-           pad_in,
-           !pad_in ? "OK (LOW as expected)" : "unexpected HIGH");
-
-    // 5. Drive HIGH again for final confirmation
-    gpio_put(PSRAM_CS_PIN, 1);
-    busy_wait_ms(1);
-    gpio0_status = *(volatile uint32_t *)0x40028000u;
-    pad_in = (gpio0_status >> 17) & 1u;
-    printf("  [SIO HIGH] INFROMPAD=%u  --> %s\n",
-           pad_in, pad_in ? "OK" : "STILL LOW");
-
-    // 6. Restore to NULL (not XIP_CS1) to prevent M1 XIP from driving CS1
-    gpio_set_function(PSRAM_CS_PIN, GPIO_FUNC_NULL);
-    printf("  [Restored NULL] GPIO0 safe — no M1 XIP interference\n");
-
-    printf("\n  Summary:\n");
-    printf("  - If SIO HIGH showed INFROMPAD=0 -> PCB short to GND (check trace/solder)\n");
-    printf("  - If SIO HIGH showed INFROMPAD=1 but XIP_CS1 is LOW -> QMI holding CS asserted\n");
-    printf("  - If ASSERT_CS1N=1 above -> QMI direct mode left CS1 stuck; clear it\n");
-}
-
-// ---------------------------------------------------------------------------
-// PSRAM SPI probe — single QMI direct mode session.
-//
-// All SPI transactions run in one RAM function to minimise time with XIP
-// paused.  The shared QSPI bus means Flash is inaccessible while CS1 is
-// asserted — each CS assertion is kept to the minimum required bytes.
-//
-// Sequence: Reset (0x66+0x99) → Read 4B → Write 4B → Read-back 4B → exit.
-// Results stored in struct, printed AFTER direct mode exits and XIP resumes.
-// ---------------------------------------------------------------------------
-struct psram_probe_result {
-    uint8_t read_id[8];     // Read ID (0x9F) response
-    uint8_t read1[8];       // SPI Read (0x03) @ 0x000000
-    uint8_t read2[4];       // read-back after write
-    bool    busy_timeout;   // true if any BUSY wait timed out
-    uint32_t gpio0_pad;     // PADS_BANK0 GPIO0 after pull-up config
-};
-
-// Wait for BUSY to clear with timeout (~1 ms). Returns true if OK.
-static bool __no_inline_not_in_flash_func(qmi_wait_busy)(void) {
-    for (uint32_t t = 125000; t; --t) {
-        if (!(qmi_hw->direct_csr & QMI_DIRECT_CSR_BUSY_BITS)) return true;
-    }
-    return false;
-}
-
-// TX one byte, OE=1, NOPUSH=1 (no RX capture). Returns true if OK.
-static bool __no_inline_not_in_flash_func(cs1_tx)(uint8_t b) {
-    qmi_hw->direct_tx = (1u << 20) | (1u << 19) | b;
-    while (!(qmi_hw->direct_csr & QMI_DIRECT_CSR_TXEMPTY_BITS));
-    return qmi_wait_busy();
-}
-
-// TX one byte in quad mode, OE=1, NOPUSH=1 (no RX capture). Returns true if OK.
-static bool __no_inline_not_in_flash_func(cs1_tx_qpi)(uint8_t b) {
-    qmi_hw->direct_tx = QMI_DIRECT_TX_NOPUSH_BITS | QMI_DIRECT_TX_OE_BITS |
-                        (QMI_DIRECT_TX_IWIDTH_VALUE_Q << QMI_DIRECT_TX_IWIDTH_LSB) | b;
-    while (!(qmi_hw->direct_csr & QMI_DIRECT_CSR_TXEMPTY_BITS));
-    return qmi_wait_busy();
-}
-
-// TX dummy, OE=1, NOPUSH=0 (capture RX). Returns true if OK.
-static bool __no_inline_not_in_flash_func(cs1_rx)(uint8_t *out) {
-    qmi_hw->direct_tx = (1u << 19) | 0x00u;
-    while (!(qmi_hw->direct_csr & QMI_DIRECT_CSR_TXEMPTY_BITS));
-    if (!qmi_wait_busy()) { *out = 0xEE; return false; }
-    *out = (uint8_t)qmi_hw->direct_rx;
-    return true;
-}
-
-// Direct register addresses for GPIO0 SIO control (accessible from RAM)
-// RP2350 SIO offsets differ from RP2040 due to interleaved GPIO_HI_* registers.
-#define GPIO0_CTRL_REG      (*(volatile uint32_t *)0x40028004u)  // IO_BANK0 GPIO0_CTRL
-#define SIO_GPIO_OUT_SET    (*(volatile uint32_t *)0xD0000018u)  // SIO + 0x18
-#define SIO_GPIO_OUT_CLR    (*(volatile uint32_t *)0xD0000020u)  // SIO + 0x20
-#define SIO_GPIO_OE_SET     (*(volatile uint32_t *)0xD0000038u)  // SIO + 0x38
-#define GPIO0_BIT           (1u << 0)
-
-// PADS_BANK0 GPIO0 register (accessible from RAM)
-#define PADS_BANK0_GPIO0    (*(volatile uint32_t *)0x40038004u)
-
-// ---------------------------------------------------------------------------
-// psram_init — full PSRAM initialisation using SIO CS + QMI clock workaround.
-//
-// Sequence (all from RAM, interrupts disabled):
-//   1. GPIO0 = SIO output HIGH + pull-up (deselect PSRAM, prevent boot glitch)
-//   2. Enter QMI direct mode, set ASSERT_CS1N (clock generation)
-//   3. SPI reset (0x66/0x99) + Read ID (0x9F) via SIO CS toggle
-//   4. Configure M1 timing/rfmt/wfmt for SPI (0x03/0x02)
-//   5. Switch GPIO0 to XIP_CS1 (FUNCSEL=9) — inside direct mode so M1 idle
-//   6. Exit direct mode → M1 XIP engine takes over GPIO0
-//   7. Enable writable on M1 window
-// ---------------------------------------------------------------------------
-struct psram_init_result {
-    uint8_t  id[8];
-    bool     id_ok;
-    bool     busy_timeout;
-    uint32_t m1_timing;
-    uint32_t m1_rfmt;
-};
-
-static void __no_inline_not_in_flash_func(psram_init_run)(
-        struct psram_init_result *r) {
-    r->busy_timeout = false;
-    r->id_ok = false;
-    bool ok = true;
-
-    // Phase 0: GPIO0 = SIO output HIGH + pull-up
-    SIO_GPIO_OUT_SET = GPIO0_BIT;
-    SIO_GPIO_OE_SET  = GPIO0_BIT;
-    GPIO0_CTRL_REG   = 5u;  // FUNCSEL = SIO
-    PADS_BANK0_GPIO0 = (1u << 6) | (1u << 4) | (1u << 3);  // IE=1 DRIVE=4mA PUE=1
-
-    for (volatile uint32_t i = 0; i < 1000; i++) __asm volatile ("nop");
-
-    // Phase 1: direct mode + ASSERT_CS1N for clock
-    uint32_t irq_save = save_and_disable_interrupts();
-    hw_set_bits(&qmi_hw->direct_csr, QMI_DIRECT_CSR_EN_BITS);
-    hw_set_bits(&qmi_hw->direct_csr, QMI_DIRECT_CSR_ASSERT_CS1N_BITS);
-
-    #define CS_INIT_LOW()  (SIO_GPIO_OUT_CLR = GPIO0_BIT)
-    #define CS_INIT_HIGH() (SIO_GPIO_OUT_SET = GPIO0_BIT)
-
-    // QPI warm-reset (device may already be in QPI mode after a warm reboot)
-    CS_INIT_LOW(); cs1_tx_qpi(0x66); CS_INIT_HIGH();
-    CS_INIT_LOW(); cs1_tx_qpi(0x99); CS_INIT_HIGH();
-    for (volatile uint32_t i = 0; i < 12500; i++) __asm volatile ("nop");
-
-    // SPI reset (bring device to known SPI state)
-    // Reset Enable (0x66)
-    CS_INIT_LOW(); ok = cs1_tx(0x66); CS_INIT_HIGH();
-    if (!ok) { r->busy_timeout = true; goto exit; }
-
-    // Reset (0x99) + tRST
-    CS_INIT_LOW(); ok = cs1_tx(0x99); CS_INIT_HIGH();
-    if (!ok) { r->busy_timeout = true; goto exit; }
-    for (volatile uint32_t i = 0; i < 12500; i++) __asm volatile ("nop");
-
-    // Read ID (0x9F + 3 addr + 8 data)
-    CS_INIT_LOW();
-    ok = cs1_tx(0x9F) && cs1_tx(0x00) && cs1_tx(0x00) && cs1_tx(0x00);
-    if (ok) {
-        for (int i = 0; i < 8; i++)
-            if (!cs1_rx(&r->id[i])) ok = false;
-    }
-    CS_INIT_HIGH();
-    if (!ok) { r->busy_timeout = true; goto exit; }
-
-    r->id_ok = (r->id[0] == 0x0D && r->id[1] == 0x5D);
-
-    // Enter QPI mode (0x35, SPI single-wire)
-    CS_INIT_LOW(); ok = cs1_tx(0x35); CS_INIT_HIGH();
-    if (!ok) { r->busy_timeout = true; goto exit; }
-
-    // Phase 2: configure M1 for QPI at 37.5 MHz (CLKDIV=2, RXDELAY=0, COOLDOWN=2)
-    // Sweep test confirmed: CLKDIV=2/RXDELAY=0 is the only zero-error setting.
-    // CLKDIV=1 (75 MHz) has ~200-300 marginal errors per 256 KB; CLKDIV=3 fails
-    // completely due to QMI sampling phase exceeding RXDELAY range.
-    // Base timing from M0 (preserves MIN_DESELECT, SELECT_HOLD etc.), then
-    // override COOLDOWN=2 (meets tCPH≥18ns), RXDELAY=0, CLKDIV=2 (SCK=37.5MHz).
-    qmi_hw->m[1].timing = (qmi_hw->m[0].timing &
-                            ~(QMI_M1_TIMING_COOLDOWN_BITS |
-                              QMI_M1_TIMING_RXDELAY_BITS  |
-                              QMI_M1_TIMING_CLKDIV_BITS))
-                         | (2u << QMI_M1_TIMING_COOLDOWN_LSB)
-                         | (0u << QMI_M1_TIMING_RXDELAY_LSB)
-                         | (2u << QMI_M1_TIMING_CLKDIV_LSB);
-
-    // rfmt: cmd=Q/8b, addr=Q, dummy=Q/28bits (7 quad clocks), data=Q
-    // APS6404L 0xEB QPI: 6 wait cycles spec; empirically requires 7 QPI clocks
-    // (likely because mode byte M[7:0] occupies 2 of the 6 wait slots in SPI mode,
-    // leaving 4+1 dummy clocks needed for QPI XIP without mode byte).
-    qmi_hw->m[1].rfmt = (QMI_M1_RFMT_PREFIX_LEN_VALUE_8  << QMI_M1_RFMT_PREFIX_LEN_LSB)  |
-                        (QMI_M1_RFMT_PREFIX_WIDTH_VALUE_Q << QMI_M1_RFMT_PREFIX_WIDTH_LSB) |
-                        (QMI_M1_RFMT_ADDR_WIDTH_VALUE_Q   << QMI_M1_RFMT_ADDR_WIDTH_LSB)   |
-                        (QMI_M1_RFMT_DUMMY_WIDTH_VALUE_Q  << QMI_M1_RFMT_DUMMY_WIDTH_LSB)  |
-                        (QMI_M1_RFMT_DATA_WIDTH_VALUE_Q   << QMI_M1_RFMT_DATA_WIDTH_LSB)   |
-                        (QMI_M1_RFMT_DUMMY_LEN_VALUE_28   << QMI_M1_RFMT_DUMMY_LEN_LSB);
-    qmi_hw->m[1].rcmd = 0xEBu;  // QPI Fast Read
-
-    // wfmt: cmd=Q/8b, addr=Q, no dummy, data=Q
-    qmi_hw->m[1].wfmt = (QMI_M1_WFMT_PREFIX_LEN_VALUE_8  << QMI_M1_WFMT_PREFIX_LEN_LSB)  |
-                        (QMI_M1_WFMT_PREFIX_WIDTH_VALUE_Q << QMI_M1_WFMT_PREFIX_WIDTH_LSB) |
-                        (QMI_M1_WFMT_ADDR_WIDTH_VALUE_Q   << QMI_M1_WFMT_ADDR_WIDTH_LSB)   |
-                        (QMI_M1_WFMT_DATA_WIDTH_VALUE_Q   << QMI_M1_WFMT_DATA_WIDTH_LSB);
-    qmi_hw->m[1].wcmd = 0x38u;  // QPI Write
-
-    r->m1_timing = qmi_hw->m[1].timing;
-    r->m1_rfmt   = qmi_hw->m[1].rfmt;
-
-    // Phase 3: switch GPIO0 to XIP_CS1 while still in direct mode
-    // M1 XIP engine is paused → won't drive GPIO0 until we exit direct mode.
-    GPIO0_CTRL_REG = 9u;  // FUNCSEL = XIP_CS1
-
-exit:
-    CS_INIT_HIGH();
-    hw_clear_bits(&qmi_hw->direct_csr, QMI_DIRECT_CSR_ASSERT_CS1N_BITS);
-    hw_clear_bits(&qmi_hw->direct_csr, QMI_DIRECT_CSR_EN_BITS);
-    restore_interrupts(irq_save);
-
-    #undef CS_INIT_LOW
-    #undef CS_INIT_HIGH
-}
-
-bool psram_init(void) {
-    struct psram_init_result r;
-    psram_init_run(&r);
-
-    if (r.id_ok)
-        hw_set_bits(&xip_ctrl_hw->ctrl, XIP_CTRL_WRITABLE_M1_BITS);
-
-    printf("[PSRAM] ID=%02X %02X -> %s  timing=0x%08X rfmt=0x%08X\n",
-           r.id[0], r.id[1],
-           r.id_ok ? "APS6404L OK" : "FAIL",
-           (unsigned)r.m1_timing, (unsigned)r.m1_rfmt);
-    if (r.busy_timeout) printf("[PSRAM] WARNING: busy timeout\n");
-
-    return r.id_ok;
-}
-
-// ---------------------------------------------------------------------------
-// psram_probe — diagnostic SPI probe (independent of psram_init).
-// ---------------------------------------------------------------------------
-static void __no_inline_not_in_flash_func(psram_probe_run)(
-        struct psram_probe_result *r) {
-    r->busy_timeout = false;
-    bool ok = true;
-
-    // --- Phase 0: GPIO0 = SIO output HIGH + internal pull-up ---
-    // GPIO0 boots with PDE=1 (pull-down) → PSRAM CE# LOW → bus contention.
-    // Fix: drive HIGH via SIO first, then enable pull-up, disable pull-down.
-    SIO_GPIO_OUT_SET = GPIO0_BIT;           // drive HIGH (CS deasserted)
-    SIO_GPIO_OE_SET  = GPIO0_BIT;           // output enable
-    GPIO0_CTRL_REG   = 5u;                  // FUNCSEL = SIO
-
-    // PADS_BANK0 GPIO0: set PUE=1, PDE=0, IE=1, OD=0
-    // Default is 0x56 = ISO=0, OD=0, IE=1, DRIVE=1, PUE=0, PDE=1, SLEWFAST=0
-    // Target:    0x4C = ISO=0, OD=0, IE=1, DRIVE=1, PUE=1, PDE=0, SLEWFAST=0
-    PADS_BANK0_GPIO0 = (1u << 6) |  // IE=1
-                       (1u << 4) |  // DRIVE=4mA (bits[5:4]=01)
-                       (1u << 3);   // PUE=1, PDE=0
-    r->gpio0_pad = PADS_BANK0_GPIO0;
-
-    // Small delay to let pull-up stabilise
-    for (volatile uint32_t i = 0; i < 1000; i++) __asm volatile ("nop");
-
-    // --- Phase 1: QMI direct mode with ASSERT_CS1N for clock generation ---
-    // We need BOTH:
-    //   SIO GPIO0 LOW/HIGH → physical CE# to APS6404L
-    //   QMI ASSERT_CS1N    → QMI generates SCLK (CS1N goes to internal pad, no ext conn)
-    uint32_t irq_save = save_and_disable_interrupts();
-    hw_set_bits(&qmi_hw->direct_csr, QMI_DIRECT_CSR_EN_BITS);
-
-    // Assert CS1N in QMI so it generates clock when we write direct_tx.
-    // This CS1N goes to internal QSPI CS1 pad (not GPIO0), harmless.
-    hw_set_bits(&qmi_hw->direct_csr, QMI_DIRECT_CSR_ASSERT_CS1N_BITS);
-
-    // CS assert/deassert via SIO (physical CE# to PSRAM)
-    #define CS_LOW()  (SIO_GPIO_OUT_CLR = GPIO0_BIT)
-    #define CS_HIGH() (SIO_GPIO_OUT_SET = GPIO0_BIT)
-
-    // --- Reset Enable (0x66) ---
-    CS_LOW();
-    ok = cs1_tx(0x66);
-    CS_HIGH();
-    if (!ok) { r->busy_timeout = true; goto exit; }
-
-    // --- Reset (0x99) ---
-    CS_LOW();
-    ok = cs1_tx(0x99);
-    CS_HIGH();
-    if (!ok) { r->busy_timeout = true; goto exit; }
-
-    // tRST ≥ 100 µs (APS6404L spec)
-    for (volatile uint32_t i = 0; i < 12500; i++) __asm volatile ("nop");
-
-    // --- Read ID (0x9F) → read_id[8] ---
-    // APS6404L: 0x9F + 3 addr bytes (0x00) + 8 data bytes
-    // Expected: MF ID = 0x0D (AP Memory), KGD = 0x5D, then EID[5:0]
-    CS_LOW();
-    ok = cs1_tx(0x9F) && cs1_tx(0x00) && cs1_tx(0x00) && cs1_tx(0x00);
-    if (ok) {
-        for (int i = 0; i < 8; i++)
-            if (!cs1_rx(&r->read_id[i])) ok = false;
-    }
-    CS_HIGH();
-    if (!ok) { r->busy_timeout = true; goto exit; }
-
-    // --- SPI Read (0x03) @ 0x000000 → read1[8] ---
-    CS_LOW();
-    ok = cs1_tx(0x03) && cs1_tx(0x00) && cs1_tx(0x00) && cs1_tx(0x00);
-    if (ok) {
-        for (int i = 0; i < 8; i++)
-            if (!cs1_rx(&r->read1[i])) ok = false;
-    }
-    CS_HIGH();
-    if (!ok) { r->busy_timeout = true; goto exit; }
-
-    // --- SPI Write (0x02) @ 0x000000 ← DE AD BE EF ---
-    CS_LOW();
-    ok = cs1_tx(0x02) && cs1_tx(0x00) && cs1_tx(0x00) && cs1_tx(0x00)
-      && cs1_tx(0xDE) && cs1_tx(0xAD) && cs1_tx(0xBE) && cs1_tx(0xEF);
-    CS_HIGH();
-    if (!ok) { r->busy_timeout = true; goto exit; }
-
-    // Short delay after write
-    for (volatile uint32_t i = 0; i < 500; i++) __asm volatile ("nop");
-
-    // --- SPI Read (0x03) @ 0x000000 → read2[4] ---
-    CS_LOW();
-    ok = cs1_tx(0x03) && cs1_tx(0x00) && cs1_tx(0x00) && cs1_tx(0x00);
-    if (ok) {
-        for (int i = 0; i < 4; i++)
-            if (!cs1_rx(&r->read2[i])) ok = false;
-    }
-    CS_HIGH();
-    if (!ok) r->busy_timeout = true;
-
-exit:
-    CS_HIGH();  // ensure physical CS deasserted
-    hw_clear_bits(&qmi_hw->direct_csr, QMI_DIRECT_CSR_ASSERT_CS1N_BITS);
-    hw_clear_bits(&qmi_hw->direct_csr, QMI_DIRECT_CSR_EN_BITS);
-    restore_interrupts(irq_save);
-
-    // Leave GPIO0 as SIO output HIGH (safe idle state, pull-up active)
-    #undef CS_LOW
-    #undef CS_HIGH
-}
-
-// ---------------------------------------------------------------------------
-// PSRAM full capacity test (APS6404L, 8 MB)
-//
-// Two-pass test: write address-based pattern across all 8 MB, then verify.
-// Reports progress every 1 MB. Requires psram_init() to have been called.
-// ---------------------------------------------------------------------------
-
-#define PSRAM_SIZE_MB    8u
-#define PSRAM_FULL_WORDS (PSRAM_SIZE_MB * 1024u * 1024u / 4u)  // 2,097,152
-#define PSRAM_STEP_WORDS (1024u * 1024u / 4u)                   // 1 MB in words (= 0x40000)
-
-void psram_full_test(void) {
-    printf("\n--- PSRAM Full Capacity Test (APS6404L, 8 MB) ---\n");
-
-    uint32_t gpio0_ctrl = *(volatile uint32_t *)0x40028004u;
-    uint32_t funcsel = gpio0_ctrl & 0x1Fu;
-    if (funcsel != 9) {
-        printf("  PSRAM not initialized (GPIO0 FUNCSEL=%u, expected 9). Run 'psram' first.\n",
-               (unsigned)funcsel);
-        return;
-    }
-
-    volatile uint32_t *psram = (volatile uint32_t *)PSRAM_NOCACHE;
-
-    // Pass 1: write address-based pattern across all 8 MB
-    printf("  Pass 1 — Write (8 MB):\n");
-    for (uint32_t i = 0; i < PSRAM_FULL_WORDS; i++) {
-        psram[i] = 0xA5000000u | i;
-        if ((i & (PSRAM_STEP_WORDS - 1u)) == 0u)
-            printf("    %u / 8 MB written\n", (unsigned)(i / PSRAM_STEP_WORDS));
-    }
-    printf("    8 / 8 MB written\n");
-
-    // Pass 2: verify all 8 MB
-    printf("  Pass 2 — Verify (8 MB):\n");
-    uint32_t errors = 0;
-    for (uint32_t i = 0; i < PSRAM_FULL_WORDS; i++) {
-        uint32_t expected = 0xA5000000u | i;
-        uint32_t actual   = psram[i];
-        if (actual != expected) {
-            if (errors < 4)
-                printf("  ERR @ +0x%06X: wr=0x%08X rd=0x%08X\n",
-                       (unsigned)(i * 4u), (unsigned)expected, (unsigned)actual);
-            errors++;
-        }
-        if ((i & (PSRAM_STEP_WORDS - 1u)) == 0u)
-            printf("    %u / 8 MB verified  errors=%u\n",
-                   (unsigned)(i / PSRAM_STEP_WORDS), (unsigned)errors);
-    }
-    printf("    8 / 8 MB verified  errors=%u\n", (unsigned)errors);
-
-    printf("\n  Capacity: %u MB (%u words)\n",
-           (unsigned)PSRAM_SIZE_MB, (unsigned)PSRAM_FULL_WORDS);
-    printf("  Errors  : %u\n", (unsigned)errors);
-    printf("  Result  : %s\n",
-           errors == 0 ? "PASS — full 8 MB accessible and data-correct"
-                       : "FAIL — data errors detected");
-}
-
-// ---------------------------------------------------------------------------
-// PSRAM XIP speed sweep — test CLKDIV × RXDELAY combinations
-//
-// For each combination, writes and reads back a test pattern to measure
-// the actual error rate.  Reports a table so the user can find the
-// fastest reliable configuration for both PSRAM and Flash.
-//
-// Two test modes per PSRAM combo:
-//   (a) Write at the candidate speed, read at the candidate speed.
-//   (b) Write at safe speed (CLKDIV=3), read at candidate speed.
-// Mode (b) isolates READ errors from WRITE errors.
-//
-// For Flash (M0): changes M0.timing from RAM with XIP paused, then
-// returns to Flash and reads a known constant.  If the fetch fails the
-// CPU faults, so this is tested last and only on user request.
-// ---------------------------------------------------------------------------
-
-#define SWEEP_WORDS  65536u   // 256 KB per pass — large enough to catch marginals
-
-// Change M1 CLKDIV + RXDELAY.  Must pause XIP via direct mode to avoid
-// corrupting any in-flight M1 transaction.  Runs from RAM.
-static void __no_inline_not_in_flash_func(psram_set_timing)(
-        uint8_t clkdiv, uint8_t rxdelay) {
-    uint32_t irq_save = save_and_disable_interrupts();
-    hw_set_bits(&qmi_hw->direct_csr, QMI_DIRECT_CSR_EN_BITS);
-
-    qmi_hw->m[1].timing = (qmi_hw->m[1].timing &
-                            ~(QMI_M1_TIMING_CLKDIV_BITS | QMI_M1_TIMING_RXDELAY_BITS))
-                         | ((uint32_t)clkdiv << QMI_M1_TIMING_CLKDIV_LSB)
-                         | ((uint32_t)rxdelay << QMI_M1_TIMING_RXDELAY_LSB);
-
-    hw_clear_bits(&qmi_hw->direct_csr, QMI_DIRECT_CSR_EN_BITS);
-    restore_interrupts(irq_save);
-}
-
-// Run a 256 KB write+verify pass.  Returns error count.
-static uint32_t psram_sweep_pass(void) {
-    volatile uint32_t *psram = (volatile uint32_t *)PSRAM_NOCACHE;
-    for (uint32_t i = 0; i < SWEEP_WORDS; i++)
-        psram[i] = 0xA5000000u | i;
-    uint32_t errors = 0;
-    for (uint32_t i = 0; i < SWEEP_WORDS; i++) {
-        if (psram[i] != (0xA5000000u | i))
-            errors++;
-    }
-    return errors;
-}
-
-// Run a 256 KB verify-only pass (data already written).  Returns error count.
-static uint32_t psram_verify_pass(void) {
-    volatile uint32_t *psram = (volatile uint32_t *)PSRAM_NOCACHE;
-    uint32_t errors = 0;
-    for (uint32_t i = 0; i < SWEEP_WORDS; i++) {
-        if (psram[i] != (0xA5000000u | i))
-            errors++;
-    }
-    return errors;
-}
-
-void psram_speed_test(void) {
-    printf("\n--- PSRAM QPI Speed Sweep (256 KB per test) ---\n");
-
-    uint32_t gpio0_ctrl = *(volatile uint32_t *)0x40028004u;
-    if ((gpio0_ctrl & 0x1Fu) != 9) {
-        printf("  PSRAM not initialized. Run 'psram' first.\n");
-        return;
-    }
-
-    // Read actual sys_clk frequency
-    uint32_t sys_hz = clock_get_hz(clk_sys);
-    printf("  sys_clk = %u MHz\n", (unsigned)(sys_hz / 1000000u));
-    printf("  Test size = %u KB (%u words)\n\n", SWEEP_WORDS * 4 / 1024, SWEEP_WORDS);
-
-    printf("  M0.timing (boot2) = 0x%08X\n", (unsigned)qmi_hw->m[0].timing);
-    printf("  M0.rfmt = 0x%08X  M0.rcmd = 0x%02X\n",
-           (unsigned)qmi_hw->m[0].rfmt, (unsigned)qmi_hw->m[0].rcmd);
-    printf("  M1.timing (init)  = 0x%08X\n", (unsigned)qmi_hw->m[1].timing);
-    printf("  M1.rfmt = 0x%08X  M1.rcmd = 0x%02X\n\n",
-           (unsigned)qmi_hw->m[1].rfmt, (unsigned)qmi_hw->m[1].rcmd);
-
-    printf("  CLKDIV  RXDELAY  SCK_MHz  M1.timing   word0_rd    wr+rd_err  rdonly_err  result\n");
-    printf("  ------  -------  -------  ----------  ----------  ---------  ----------  ------\n");
-
-    static const uint8_t clkdivs[]  = {3, 2, 1};
-    static const uint8_t rxdelays[] = {0, 1, 2, 3, 4, 5, 6, 7};
-
-    uint32_t orig_timing = qmi_hw->m[1].timing;
-
-    for (int c = 0; c < 3; c++) {
-        for (int r = 0; r < 8; r++) {
-            uint8_t cd = clkdivs[c];
-            uint8_t rd = rxdelays[r];
-            uint32_t sck_mhz = sys_hz / (2u * cd) / 1000000u;
-
-            // (a) write+read at candidate speed
-            psram_set_timing(cd, rd);
-            uint32_t tim_val = qmi_hw->m[1].timing;
-            volatile uint32_t *psram = (volatile uint32_t *)PSRAM_NOCACHE;
-            psram[0] = 0xDEADBEEFu;
-            uint32_t word0 = psram[0];
-            uint32_t wr_err = psram_sweep_pass();
-
-            // (b) write at safe speed, read at candidate speed
-            psram_set_timing(2, 0);                   // safe: 37.5 MHz, RXDELAY=0
-            for (uint32_t i = 0; i < SWEEP_WORDS; i++)
-                psram[i] = 0xA5000000u | i;
-            psram_set_timing(cd, rd);                  // switch to candidate
-            uint32_t rd_err = psram_verify_pass();
-
-            bool pass = (wr_err == 0) && (rd_err == 0);
-            printf("  %6u  %7u  %7u  0x%08X  0x%08X  %9u  %10u  %s\n",
-                   cd, rd, sck_mhz, (unsigned)tim_val, (unsigned)word0,
-                   (unsigned)wr_err, (unsigned)rd_err,
-                   pass ? "PASS" : "FAIL");
-        }
-    }
-
-    // Restore original timing
-    qmi_hw->m[1].timing = orig_timing;
-    printf("\n  M1.timing restored to 0x%08X\n", (unsigned)orig_timing);
-    printf("  Current: CLKDIV=%u RXDELAY=%u SCK=%u MHz\n",
-           (unsigned)(orig_timing & 0xFF),
-           (unsigned)((orig_timing >> 8) & 0x7),
-           (unsigned)(sys_hz / (2u * (orig_timing & 0xFF)) / 1000000u));
-}
-
-// ---------------------------------------------------------------------------
-// PSRAM QPI CLKDIV=1 diagnostic — analyse error patterns and timing tuning.
-//
-// Tests:
-//  1) Error address analysis at CLKDIV=1/RXDELAY=0 — page boundary vs random
-//  2) Sweep COOLDOWN (0-3) × MIN_DESELECT (1,3,5,7) at CLKDIV=1/RXDELAY=0
-//  3) Sweep PAGEBREAK (0-3) at CLKDIV=1/RXDELAY=0
-//  4) Multiple runs to check error reproducibility
-// ---------------------------------------------------------------------------
-
-// Change M1 timing with full field control.  Runs from RAM.
-static void __no_inline_not_in_flash_func(psram_set_full_timing)(
-        uint32_t timing) {
-    uint32_t irq_save = save_and_disable_interrupts();
-    hw_set_bits(&qmi_hw->direct_csr, QMI_DIRECT_CSR_EN_BITS);
-    qmi_hw->m[1].timing = timing;
-    hw_clear_bits(&qmi_hw->direct_csr, QMI_DIRECT_CSR_EN_BITS);
-    restore_interrupts(irq_save);
-}
-
-void psram_diag_test(void) {
-    printf("\n--- PSRAM QPI CLKDIV=1 Diagnostic ---\n");
-
-    uint32_t gpio0_ctrl = *(volatile uint32_t *)0x40028004u;
-    if ((gpio0_ctrl & 0x1Fu) != 9) {
-        printf("  PSRAM not initialized. Run 'psram' first.\n");
-        return;
-    }
-
-    uint32_t sys_hz = clock_get_hz(clk_sys);
-    uint32_t orig_timing = qmi_hw->m[1].timing;
-    volatile uint32_t *psram = (volatile uint32_t *)PSRAM_NOCACHE;
-
-    // ---- Test 1: RXDELAY=0 (10 runs, write@safe read@75MHz) ----
-    printf("\n  [1] CLKDIV=1/RXDELAY=0 (75 MHz), 10 runs\n");
-    for (int run = 0; run < 10; run++) {
-        psram_set_timing(2, 0);
-        for (uint32_t i = 0; i < SWEEP_WORDS; i++)
-            psram[i] = 0xA5000000u | i;
-        psram_set_timing(1, 0);
-        uint32_t errs = psram_verify_pass();
-        printf("      Run %2d: %u errors\n", run + 1, (unsigned)errs);
-    }
-
-    // ---- Test 2: RXDELAY=1 (10 runs, write@safe read@75MHz) ----
-    printf("\n  [2] CLKDIV=1/RXDELAY=1 (75 MHz), 10 runs\n");
-    for (int run = 0; run < 10; run++) {
-        psram_set_timing(2, 0);
-        for (uint32_t i = 0; i < SWEEP_WORDS; i++)
-            psram[i] = 0xA5000000u | i;
-        psram_set_timing(1, 1);
-        uint32_t errs = psram_verify_pass();
-        printf("      Run %2d: %u errors\n", run + 1, (unsigned)errs);
-    }
-
-    // ---- Test 3: CLKDIV=2/RXDELAY=0 baseline (5 runs for reference) ----
-    printf("\n  [3] CLKDIV=2/RXDELAY=0 (37.5 MHz) baseline, 5 runs\n");
-    for (int run = 0; run < 5; run++) {
-        psram_set_timing(2, 0);
-        for (uint32_t i = 0; i < SWEEP_WORDS; i++)
-            psram[i] = 0xA5000000u | i;
-        uint32_t errs = psram_verify_pass();
-        printf("      Run %2d: %u errors\n", run + 1, (unsigned)errs);
-    }
-
-    // ---- Test 4: CLKDIV=1/RXDELAY=0, write+read both at 75 MHz (5 runs) ----
-    printf("\n  [4] CLKDIV=1/RXDELAY=0 write+read both at 75 MHz, 5 runs\n");
-    for (int run = 0; run < 5; run++) {
-        psram_set_timing(1, 0);
-        uint32_t errs = psram_sweep_pass();
-        printf("      Run %2d: %u errors\n", run + 1, (unsigned)errs);
-    }
-
-    // Restore original timing
-    psram_set_full_timing(orig_timing);
-    printf("\n  Restored M1.timing = 0x%08X\n", (unsigned)orig_timing);
-}
-
-// ---------------------------------------------------------------------------
-// Flash (M0) XIP speed test — change M0 timing from RAM, test Flash reads.
-//
-// This function runs entirely from RAM (no_inline_not_in_flash) so that XIP
-// can be safely paused while M0.timing is changed.  It reads back a known
-// Flash constant after the change to verify correctness.
-// ---------------------------------------------------------------------------
-
-// Known sentinel in Flash — address and value set by the linker.
-// We use the first word of the vector table (Initial MSP) which is always
-// 0x20082000 for RP2350B with 520 KB SRAM.
-#define FLASH_TEST_ADDR   0x10000000u
-#define FLASH_TEST_EXPECT (*(volatile uint32_t *)FLASH_TEST_ADDR)
-
-struct flash_speed_result {
-    uint8_t  clkdiv;
-    uint8_t  rxdelay;
-    uint32_t sck_mhz;
-    uint32_t read_val;
-    uint32_t expected;
-    bool     pass;
-};
-
-static void __no_inline_not_in_flash_func(flash_speed_run)(
-        struct flash_speed_result *results, int count,
-        const uint8_t *clkdivs, const uint8_t *rxdelays,
-        uint32_t sys_hz) {
-    // Read expected value BEFORE changing timing (while XIP still works)
-    uint32_t expected = *(volatile uint32_t *)FLASH_TEST_ADDR;
-
-    for (int i = 0; i < count; i++) {
-        uint8_t cd = clkdivs[i];
-        uint8_t rd = rxdelays[i];
-
-        // Pause XIP
-        uint32_t irq_save = save_and_disable_interrupts();
-        hw_set_bits(&qmi_hw->direct_csr, QMI_DIRECT_CSR_EN_BITS);
-
-        // Save and change M0 timing
-        uint32_t orig = qmi_hw->m[0].timing;
-        qmi_hw->m[0].timing = (orig & ~(QMI_M0_TIMING_CLKDIV_BITS |
-                                         QMI_M0_TIMING_RXDELAY_BITS))
-                             | ((uint32_t)cd << QMI_M0_TIMING_CLKDIV_LSB)
-                             | ((uint32_t)rd << QMI_M0_TIMING_RXDELAY_LSB);
-
-        // Resume XIP with new timing
-        hw_clear_bits(&qmi_hw->direct_csr, QMI_DIRECT_CSR_EN_BITS);
-        restore_interrupts(irq_save);
-
-        // Read test word from Flash (uses new timing)
-        uint32_t val = *(volatile uint32_t *)FLASH_TEST_ADDR;
-
-        // Revert to original timing immediately (from RAM — safe even if
-        // the read above returned garbage, because we are still in RAM)
-        irq_save = save_and_disable_interrupts();
-        hw_set_bits(&qmi_hw->direct_csr, QMI_DIRECT_CSR_EN_BITS);
-        qmi_hw->m[0].timing = orig;
-        hw_clear_bits(&qmi_hw->direct_csr, QMI_DIRECT_CSR_EN_BITS);
-        restore_interrupts(irq_save);
-
-        results[i].clkdiv  = cd;
-        results[i].rxdelay = rd;
-        results[i].sck_mhz = sys_hz / (2u * cd) / 1000000u;
-        results[i].read_val = val;
-        results[i].expected = expected;
-        results[i].pass     = (val == expected);
-    }
-}
-
-void flash_speed_test(void) {
-    printf("\n--- Flash (M0) XIP Speed Test ---\n");
-
-    uint32_t sys_hz = clock_get_hz(clk_sys);
-    printf("  sys_clk = %u MHz\n", (unsigned)(sys_hz / 1000000u));
-    printf("  Test: read vector table MSP @ 0x%08X\n", FLASH_TEST_ADDR);
-    printf("  Expected value: 0x%08X\n\n", (unsigned)FLASH_TEST_EXPECT);
-
-    // Build test matrix: CLKDIV=1..4 × RXDELAY=0..3
-    #define FLASH_COMBOS 16
-    uint8_t cd_arr[FLASH_COMBOS], rd_arr[FLASH_COMBOS];
-    struct flash_speed_result res[FLASH_COMBOS];
-    int idx = 0;
-    for (uint8_t cd = 1; cd <= 4; cd++)
-        for (uint8_t rd = 0; rd <= 3; rd++) {
-            cd_arr[idx] = cd;
-            rd_arr[idx] = rd;
-            idx++;
-        }
-
-    flash_speed_run(res, FLASH_COMBOS, cd_arr, rd_arr, sys_hz);
-
-    printf("  CLKDIV  RXDELAY  SCK_MHz  read_val    result\n");
-    printf("  ------  -------  -------  ----------  ------\n");
-    for (int i = 0; i < FLASH_COMBOS; i++) {
-        printf("  %6u  %7u  %7u  0x%08X  %s\n",
-               res[i].clkdiv, res[i].rxdelay, res[i].sck_mhz,
-               (unsigned)res[i].read_val,
-               res[i].pass ? "PASS" : "FAIL");
-    }
-    #undef FLASH_COMBOS
-}
-
-// ---------------------------------------------------------------------------
-// PSRAM J-Link probe prep
-//
-// Writes four sentinel words to PSRAM[0..3], reads them back, then prints the
-// J-Link Commander commands needed to verify PSRAM is accessible via SWD
-// while the MCU is halted.
-// ---------------------------------------------------------------------------
-
-void psram_jlink_prep(void) {
-    printf("\n--- PSRAM J-Link access probe ---\n");
-
-    uint32_t gpio0_ctrl = *(volatile uint32_t *)0x40028004u;
-    if ((gpio0_ctrl & 0x1Fu) != 9) {
-        printf("  PSRAM not initialized (GPIO0 FUNCSEL=%u). Run 'psram' first.\n",
-               (unsigned)(gpio0_ctrl & 0x1Fu));
-        return;
-    }
-
-    volatile uint32_t *psram = (volatile uint32_t *)PSRAM_NOCACHE;
-
-    static const uint32_t sentinel[4] = {
-        0xDEADBEEFu, 0xA5A5A5A5u, 0xCAFEBABEu, 0x12345678u
-    };
-    for (int i = 0; i < 4; i++) psram[i] = sentinel[i];
-
-    uint32_t rb[4];
-    for (int i = 0; i < 4; i++) rb[i] = psram[i];
-
-    bool all_ok = true;
-    printf("  XIP uncached base : 0x%08X\n", (unsigned)PSRAM_NOCACHE);
-    printf("  Sentinel write+readback:\n");
-    for (int i = 0; i < 4; i++) {
-        bool ok = (rb[i] == sentinel[i]);
-        if (!ok) all_ok = false;
-        printf("    [%d] wr=0x%08X  rd=0x%08X  %s\n", i,
-               (unsigned)sentinel[i], (unsigned)rb[i], ok ? "OK" : "MISMATCH");
-    }
-    printf("  Firmware readback: %s\n\n", all_ok ? "PASS" : "FAIL");
-
-    printf("  J-Link Commander commands (halt MCU, read, resume):\n");
-    printf("    connect\n");
-    printf("    h\n");
-    printf("    mem32 0x%08X 4\n", (unsigned)PSRAM_NOCACHE);
-    printf("    Expected: %08X %08X %08X %08X\n",
-           (unsigned)sentinel[0], (unsigned)sentinel[1],
-           (unsigned)sentinel[2], (unsigned)sentinel[3]);
-    printf("    g\n");
-    printf("    qc\n");
-}
-
-void psram_probe(void) {
-    printf("\n--- PSRAM SPI probe (SIO CS + QMI clock, GPIO0) ---\n");
-
-    struct psram_probe_result r;
-    psram_probe_run(&r);
-
-    // Print results (safe — direct mode exited, XIP active)
-    printf("  GPIO0 pad: 0x%08X (PUE=%u PDE=%u IE=%u)\n",
-           (unsigned)r.gpio0_pad,
-           (r.gpio0_pad >> 3) & 1, (r.gpio0_pad >> 2) & 1,
-           (r.gpio0_pad >> 6) & 1);
-
-    printf("  Read ID:   %02X %02X %02X %02X %02X %02X %02X %02X",
-           r.read_id[0], r.read_id[1], r.read_id[2], r.read_id[3],
-           r.read_id[4], r.read_id[5], r.read_id[6], r.read_id[7]);
-    bool id_ok = (r.read_id[0] == 0x0D && r.read_id[1] == 0x5D);
-    printf("  %s\n", id_ok ? "(APS6404L OK: MF=0x0D KGD=0x5D)" :
-                             "(unexpected — check wiring)");
-
-    printf("  SPI Read:  %02X %02X %02X %02X %02X %02X %02X %02X\n",
-           r.read1[0], r.read1[1], r.read1[2], r.read1[3],
-           r.read1[4], r.read1[5], r.read1[6], r.read1[7]);
-
-    printf("  Write:     DE AD BE EF\n");
-    printf("  Read-back: %02X %02X %02X %02X",
-           r.read2[0], r.read2[1], r.read2[2], r.read2[3]);
-    bool match = (r.read2[0]==0xDE && r.read2[1]==0xAD &&
-                  r.read2[2]==0xBE && r.read2[3]==0xEF);
-    printf("  %s\n", match ? "PASS" : "MISMATCH");
-
-    if (r.busy_timeout) printf("  WARNING: BUSY timeout occurred\n");
-
-    printf("  M1: timing=0x%08X rfmt=0x%08X rcmd=0x%08X\n",
-           (unsigned)qmi_hw->m[1].timing, (unsigned)qmi_hw->m[1].rfmt,
-           (unsigned)qmi_hw->m[1].rcmd);
-}
-
-// ===========================================================================
-// TFT consolidated diagnostics (Step 20)
-// ===========================================================================
-
-// ---------------------------------------------------------------------------
-// Memory Diagnostic — consolidated SRAM + Flash + PSRAM on one TFT screen
-//
-// Sub-tests:
-//   1. SRAM 16 KB 5-pattern test
-//   2. Flash JEDEC ID (W25Q128JW: EF 60 18)
-//   3. Flash QE bit (SR2 bit 1)
-//   4. PSRAM init (APS6404L ID 0D 5D)
-//   5. PSRAM 4 KB pattern test
-// ---------------------------------------------------------------------------
-void cmd_memory_diag(void) {
-    const int S = 2;
-    const int CW = 6 * S;
-    const int CH = 8 * S;
-    const int W  = menu_tft_width();
-    const int COLS = W / CW;
-
-    menu_clear(MC_BG);
-    menu_str(0, 0, " Memory Diagnostic  ", COLS, MC_TITLE, MC_TITBG, S);
-
-    printf("\n--- Memory Diagnostic (consolidated) ---\n");
-    int pass = 0;
-    int total = 5;
-    char line[24];
-
-    // --- 1. SRAM 5-pattern test ---
-    {
-        static uint32_t buf[4096]; // 16 KB
-        static const uint32_t pats[] = {
-            0xAAAAAAAAu, 0x55555555u, 0xFF00FF00u, 0x00FF00FFu
-        };
-        int errs = 0;
-        for (int p = 0; p < 4; p++) {
-            for (int i = 0; i < 4096; i++) buf[i] = pats[p];
-            for (int i = 0; i < 4096; i++)
-                if (buf[i] != pats[p]) errs++;
-        }
-        // Address pattern
-        for (int i = 0; i < 4096; i++) buf[i] = 0xA5000000u | (uint32_t)i;
-        for (int i = 0; i < 4096; i++)
-            if (buf[i] != (0xA5000000u | (uint32_t)i)) errs++;
-
-        bool ok = (errs == 0);
-        if (ok) pass++;
-        snprintf(line, sizeof(line), " SRAM 16K 5pat  %s", ok ? " OK" : "ERR");
-        menu_str(0, 2 * CH, line, COLS, ok ? MC_OK : MC_ERR, MC_BG, S);
-        printf("  SRAM 16KB 5-pattern: %s (%d errors)\n", ok ? "PASS" : "FAIL", errs);
-    }
-
-    // --- 2. Flash JEDEC ID ---
-    {
-        uint8_t txbuf[4] = {0x9F, 0, 0, 0};
-        uint8_t rxbuf[4] = {0};
-        flash_do_cmd(txbuf, rxbuf, 4);
-        uint8_t mfr = rxbuf[1], type = rxbuf[2], cap = rxbuf[3];
-        bool ok = (mfr == 0xEF) && (type == 0x60) && (cap == 0x18);
-        if (ok) pass++;
-        snprintf(line, sizeof(line), " Flash %02X%02X%02X   %s",
-                 mfr, type, cap, ok ? " OK" : "ERR");
-        menu_str(0, 3 * CH, line, COLS, ok ? MC_OK : MC_ERR, MC_BG, S);
-        printf("  Flash JEDEC %02X %02X %02X: %s\n", mfr, type, cap,
-               ok ? "PASS" : "FAIL");
-    }
-
-    // --- 3. Flash QE bit ---
-    {
-        uint8_t sr_tx[2], sr_rx[2];
-        sr_tx[0] = 0x35; sr_tx[1] = 0;
-        flash_do_cmd(sr_tx, sr_rx, 2);
-        bool qe = (sr_rx[1] & 0x02) != 0;
-        if (qe) pass++;
-        snprintf(line, sizeof(line), " Flash QE=%d SR2 %s",
-                 qe ? 1 : 0, qe ? " OK" : "ERR");
-        menu_str(0, 4 * CH, line, COLS, qe ? MC_OK : MC_ERR, MC_BG, S);
-        printf("  Flash QE (SR2 bit1): %s\n", qe ? "SET (OK)" : "CLEAR (FAIL)");
-    }
-
-    // --- 4. PSRAM init ---
-    {
-        bool ok = psram_init();
-        if (ok) pass++;
-        snprintf(line, sizeof(line), " PSRAM Init QPI %s", ok ? " OK" : "ERR");
-        menu_str(0, 5 * CH, line, COLS, ok ? MC_OK : MC_ERR, MC_BG, S);
-        printf("  PSRAM Init: %s\n", ok ? "PASS" : "FAIL");
-
-        // --- 5. PSRAM 4 KB pattern (only if init succeeded) ---
-        if (ok) {
-            volatile uint32_t *pm = (volatile uint32_t *)PSRAM_NOCACHE;
-            for (int i = 0; i < 1024; i++) pm[i] = 0xA5000000u | (uint32_t)i;
-            int errs = 0;
-            for (int i = 0; i < 1024; i++)
-                if (pm[i] != (0xA5000000u | (uint32_t)i)) errs++;
-            bool pok = (errs == 0);
-            if (pok) pass++;
-            snprintf(line, sizeof(line), " PSRAM 4K pat  %s", pok ? " OK" : "ERR");
-            menu_str(0, 6 * CH, line, COLS, pok ? MC_OK : MC_ERR, MC_BG, S);
-            printf("  PSRAM 4KB pattern: %s (%d errors)\n",
-                   pok ? "PASS" : "FAIL", errs);
-        } else {
-            menu_str(0, 6 * CH, " PSRAM 4K pat  SKIP", COLS, MC_HINT, MC_BG, S);
-            printf("  PSRAM 4KB pattern: SKIP (init failed)\n");
-        }
-    }
-
-    // --- Summary ---
-    snprintf(line, sizeof(line), " Result: %d/%d pass", pass, total);
-    menu_str(0, 8 * CH, line, COLS, pass == total ? MC_OK : MC_ERR, MC_BG, S);
-    menu_str(0, 10 * CH, " BACK to return     ", COLS, MC_HINT, MC_BG, S);
-
-    printf("\n  === %d/%d tests OK ===\n", pass, total);
-
-    while (!back_key_pressed()) sleep_ms(50);
-}
-
-// ---------------------------------------------------------------------------
-// PSRAM Full 8 MB test with TFT progress display
-// ---------------------------------------------------------------------------
-void cmd_psram_full_tft(void) {
-    const int S = 2;
-    const int CW = 6 * S;
-    const int CH = 8 * S;
-    const int W  = menu_tft_width();
-    const int COLS = W / CW;
-
-    menu_clear(MC_BG);
-    menu_str(0, 0, " PSRAM Full Test    ", COLS, MC_TITLE, MC_TITBG, S);
-
-    printf("\n--- PSRAM Full 8 MB Test (TFT) ---\n");
-
-    // Check PSRAM init
-    uint32_t gpio0_ctrl = *(volatile uint32_t *)0x40028004u;
-    if ((gpio0_ctrl & 0x1Fu) != 9) {
-        menu_str(0, 2 * CH, " PSRAM not init'd   ", COLS, MC_ERR, MC_BG, S);
-        menu_str(0, 3 * CH, " Run MemDiag first  ", COLS, MC_HINT, MC_BG, S);
-        menu_str(0, 10 * CH, " BACK to return     ", COLS, MC_HINT, MC_BG, S);
-        printf("  PSRAM not initialized.\n");
-        while (!back_key_pressed()) sleep_ms(50);
-        return;
-    }
-
-    volatile uint32_t *psram = (volatile uint32_t *)PSRAM_NOCACHE;
-    char line[24];
-
-    // Pass 1: Write
-    menu_str(0, 2 * CH, " Writing...         ", COLS, MC_FG, MC_BG, S);
-    printf("  Pass 1 — Write (8 MB):\n");
-    for (uint32_t mb = 0; mb < PSRAM_SIZE_MB; mb++) {
-        uint32_t base = mb * PSRAM_STEP_WORDS;
-        for (uint32_t i = 0; i < PSRAM_STEP_WORDS; i++)
-            psram[base + i] = 0xA5000000u | (base + i);
-
-        snprintf(line, sizeof(line), " Write: %u/8 MB     ", (unsigned)(mb + 1));
-        menu_str(0, 3 * CH, line, COLS, MC_FG, MC_BG, S);
-        printf("    %u / 8 MB written\n", (unsigned)(mb + 1));
-
-        if (back_key_pressed()) {
-            menu_str(0, 8 * CH, " Cancelled          ", COLS, MC_HINT, MC_BG, S);
-            menu_str(0, 10 * CH, " BACK to return     ", COLS, MC_HINT, MC_BG, S);
-            while (!back_key_pressed()) sleep_ms(50);
-            return;
-        }
-    }
-
-    // Pass 2: Verify
-    menu_str(0, 5 * CH, " Verifying...       ", COLS, MC_FG, MC_BG, S);
-    printf("  Pass 2 — Verify (8 MB):\n");
-    uint32_t errors = 0;
-    for (uint32_t mb = 0; mb < PSRAM_SIZE_MB; mb++) {
-        uint32_t base = mb * PSRAM_STEP_WORDS;
-        for (uint32_t i = 0; i < PSRAM_STEP_WORDS; i++) {
-            uint32_t expected = 0xA5000000u | (base + i);
-            if (psram[base + i] != expected) errors++;
-        }
-
-        snprintf(line, sizeof(line), " Vfy: %u/8 err=%u",
-                 (unsigned)(mb + 1), (unsigned)errors);
-        menu_str(0, 6 * CH, line, COLS, MC_FG, MC_BG, S);
-        printf("    %u / 8 MB verified  errors=%u\n",
-               (unsigned)(mb + 1), (unsigned)errors);
-
-        if (back_key_pressed()) {
-            menu_str(0, 8 * CH, " Cancelled          ", COLS, MC_HINT, MC_BG, S);
-            menu_str(0, 10 * CH, " BACK to return     ", COLS, MC_HINT, MC_BG, S);
-            while (!back_key_pressed()) sleep_ms(50);
-            return;
-        }
-    }
-
-    // Result
-    bool ok = (errors == 0);
-    snprintf(line, sizeof(line), " Errors: %u         ", (unsigned)errors);
-    menu_str(0, 7 * CH, line, COLS, MC_FG, MC_BG, S);
-    snprintf(line, sizeof(line), " Result: %s        ", ok ? "PASS" : "FAIL");
-    menu_str(0, 8 * CH, line, COLS, ok ? MC_OK : MC_ERR, MC_BG, S);
-    menu_str(0, 10 * CH, " BACK to return     ", COLS, MC_HINT, MC_BG, S);
-
-    printf("\n  Result: %s (%u errors)\n", ok ? "PASS" : "FAIL", (unsigned)errors);
-
-    while (!back_key_pressed()) sleep_ms(50);
-}
-
-// ---------------------------------------------------------------------------
-// PSRAM Tuning — merged sweep + diag + flash_sweep, TFT summary
-//
-// Runs PSRAM speed sweep (CLKDIV 1-3 × RXDELAY 0-7) and Flash speed sweep
-// (CLKDIV 1-4 × RXDELAY 0-3).  Displays best-case results on TFT.
-// Full sweep table is printed to serial.
-// ---------------------------------------------------------------------------
-void cmd_psram_tuning(void) {
-    const int S = 2;
-    const int CW = 6 * S;
-    const int CH = 8 * S;
-    const int W  = menu_tft_width();
-    const int COLS = W / CW;
-
-    menu_clear(MC_BG);
-    menu_str(0, 0, " PSRAM/Flash Tuning ", COLS, MC_TITLE, MC_TITBG, S);
-
-    // Check PSRAM init
-    uint32_t gpio0_ctrl = *(volatile uint32_t *)0x40028004u;
-    if ((gpio0_ctrl & 0x1Fu) != 9) {
-        menu_str(0, 2 * CH, " PSRAM not init'd   ", COLS, MC_ERR, MC_BG, S);
-        menu_str(0, 3 * CH, " Run MemDiag first  ", COLS, MC_HINT, MC_BG, S);
-        menu_str(0, 10 * CH, " BACK to return     ", COLS, MC_HINT, MC_BG, S);
-        printf("  PSRAM not initialized.\n");
-        while (!back_key_pressed()) sleep_ms(50);
-        return;
-    }
-
-    uint32_t sys_hz = clock_get_hz(clk_sys);
-    uint32_t orig_timing = qmi_hw->m[1].timing;
-    char line[24];
-
-    menu_str(0, 2 * CH, " PSRAM sweep...     ", COLS, MC_FG, MC_BG, S);
-    printf("\n--- PSRAM/Flash Tuning (TFT) ---\n");
-    printf("  sys_clk = %u MHz\n", (unsigned)(sys_hz / 1000000u));
-
-    // --- PSRAM sweep ---
-    static const uint8_t clkdivs[]  = {3, 2, 1};
-    static const uint8_t rxdelays[] = {0, 1, 2, 3, 4, 5, 6, 7};
-
-    printf("\n  PSRAM Sweep (256 KB per test):\n");
-    printf("  CLKDIV  RXDELAY  SCK_MHz  wr+rd_err  rdonly_err  result\n");
-    printf("  ------  -------  -------  ---------  ----------  ------\n");
-
-    uint8_t best_cd = 0;
-    uint8_t best_rd = 0;
-    uint32_t best_mhz = 0;
-    int psram_pass_count = 0;
-
-    for (int c = 0; c < 3; c++) {
-        for (int r = 0; r < 8; r++) {
-            uint8_t cd = clkdivs[c];
-            uint8_t rd = rxdelays[r];
-            uint32_t sck_mhz = sys_hz / (2u * cd) / 1000000u;
-
-            // Write+read at candidate speed
-            psram_set_timing(cd, rd);
-            uint32_t wr_err = psram_sweep_pass();
-
-            // Write at safe speed, read at candidate speed
-            psram_set_timing(2, 0);
-            volatile uint32_t *pm = (volatile uint32_t *)PSRAM_NOCACHE;
-            for (uint32_t i = 0; i < SWEEP_WORDS; i++)
-                pm[i] = 0xA5000000u | i;
-            psram_set_timing(cd, rd);
-            uint32_t rd_err = psram_verify_pass();
-
-            bool ok = (wr_err == 0) && (rd_err == 0);
-            if (ok) {
-                psram_pass_count++;
-                if (sck_mhz > best_mhz) {
-                    best_mhz = sck_mhz;
-                    best_cd = cd;
-                    best_rd = rd;
-                }
-            }
-
-            printf("  %6u  %7u  %7u  %9u  %10u  %s\n",
-                   cd, rd, sck_mhz, (unsigned)wr_err, (unsigned)rd_err,
-                   ok ? "PASS" : "FAIL");
-
-            if (back_key_pressed()) goto tuning_done;
-        }
-    }
-
-    // Restore PSRAM timing
-    psram_set_full_timing(orig_timing);
-
-    // Display PSRAM results on TFT
-    if (best_mhz > 0) {
-        snprintf(line, sizeof(line), " Best: CD%u RD%u %uMHz",
-                 best_cd, best_rd, (unsigned)best_mhz);
-        menu_str(0, 3 * CH, line, COLS, MC_OK, MC_BG, S);
-    } else {
-        menu_str(0, 3 * CH, " No PASS combo      ", COLS, MC_ERR, MC_BG, S);
-    }
-    snprintf(line, sizeof(line), " Pass: %d/24 combos ", psram_pass_count);
-    menu_str(0, 4 * CH, line, COLS, MC_FG, MC_BG, S);
-
-    // --- Flash sweep ---
-    menu_str(0, 6 * CH, " Flash sweep...     ", COLS, MC_FG, MC_BG, S);
-    printf("\n  Flash (M0) Speed Sweep:\n");
-
-    #define TUNE_FLASH_COMBOS 16
-    uint8_t cd_arr[TUNE_FLASH_COMBOS], rd_arr[TUNE_FLASH_COMBOS];
-    struct flash_speed_result fres[TUNE_FLASH_COMBOS];
-    int idx = 0;
-    for (uint8_t cd = 1; cd <= 4; cd++)
-        for (uint8_t rd = 0; rd <= 3; rd++) {
-            cd_arr[idx] = cd;
-            rd_arr[idx] = rd;
-            idx++;
-        }
-    flash_speed_run(fres, TUNE_FLASH_COMBOS, cd_arr, rd_arr, sys_hz);
-
-    printf("  CLKDIV  RXDELAY  SCK_MHz  read_val    result\n");
-    int flash_pass_count = 0;
-    uint32_t flash_best_mhz = 0;
-    uint8_t flash_best_cd = 0, flash_best_rd = 0;
-    for (int i = 0; i < TUNE_FLASH_COMBOS; i++) {
-        printf("  %6u  %7u  %7u  0x%08X  %s\n",
-               fres[i].clkdiv, fres[i].rxdelay, fres[i].sck_mhz,
-               (unsigned)fres[i].read_val,
-               fres[i].pass ? "PASS" : "FAIL");
-        if (fres[i].pass) {
-            flash_pass_count++;
-            if (fres[i].sck_mhz > flash_best_mhz) {
-                flash_best_mhz = fres[i].sck_mhz;
-                flash_best_cd = fres[i].clkdiv;
-                flash_best_rd = fres[i].rxdelay;
-            }
-        }
-    }
-    #undef TUNE_FLASH_COMBOS
-
-    if (flash_best_mhz > 0) {
-        snprintf(line, sizeof(line), " Best: CD%u RD%u %uMHz",
-                 flash_best_cd, flash_best_rd, (unsigned)flash_best_mhz);
-        menu_str(0, 7 * CH, line, COLS, MC_OK, MC_BG, S);
-    } else {
-        menu_str(0, 7 * CH, " No PASS combo      ", COLS, MC_ERR, MC_BG, S);
-    }
-    snprintf(line, sizeof(line), " Pass: %d/16 combos ", flash_pass_count);
-    menu_str(0, 8 * CH, line, COLS, MC_FG, MC_BG, S);
-
-tuning_done:
-    psram_set_full_timing(orig_timing);
-    menu_str(0, 10 * CH, " BACK to return     ", COLS, MC_HINT, MC_BG, S);
-    printf("\n  M1.timing restored to 0x%08X\n", (unsigned)orig_timing);
-
-    while (!back_key_pressed()) sleep_ms(50);
-}
-
-// ---------------------------------------------------------------------------
-// PSRAM Debug — init + XIP sentinel + M1 timing info, TFT display
-//
-// Sub-tests:
-//   1. PSRAM Init (psram_init: SPI reset, Read ID, QPI, XIP config)
-//   2. XIP Sentinel (4 words write+readback via memory-mapped XIP)
-// Low-level SPI probe is available via serial command 'psram_probe'.
-// ---------------------------------------------------------------------------
-void cmd_psram_debug(void) {
-    const int S = 2;
-    const int CW = 6 * S;
-    const int CH = 8 * S;
-    const int W  = menu_tft_width();
-    const int COLS = W / CW;
-
-    menu_clear(MC_BG);
-    menu_str(0, 0, " PSRAM Debug        ", COLS, MC_TITLE, MC_TITBG, S);
-
-    printf("\n--- PSRAM Debug (init + sentinel) ---\n");
-    char line[24];
-    int pass = 0;
-    int total = 2;
-
-    // --- 1. PSRAM Init ---
-    menu_str(0, 2 * CH, " Initializing...    ", COLS, MC_FG, MC_BG, S);
-
-    bool init_ok = psram_init();
-    if (init_ok) pass++;
-    snprintf(line, sizeof(line), " Init QPI      %s", init_ok ? " OK" : "ERR");
-    menu_str(0, 2 * CH, line, COLS, init_ok ? MC_OK : MC_ERR, MC_BG, S);
-    printf("  PSRAM Init: %s\n", init_ok ? "PASS" : "FAIL");
-
-    // --- 2. XIP Sentinel ---
-    if (init_ok) {
-        volatile uint32_t *pm = (volatile uint32_t *)PSRAM_NOCACHE;
-        static const uint32_t sentinel[4] = {
-            0xDEADBEEFu, 0xA5A5A5A5u, 0xCAFEBABEu, 0x12345678u
-        };
-        for (int i = 0; i < 4; i++) pm[i] = sentinel[i];
-
-        bool all_ok = true;
-        for (int i = 0; i < 4; i++) {
-            if (pm[i] != sentinel[i]) all_ok = false;
-        }
-        if (all_ok) pass++;
-
-        snprintf(line, sizeof(line), " XIP sentinel  %s", all_ok ? " OK" : "ERR");
-        menu_str(0, 3 * CH, line, COLS, all_ok ? MC_OK : MC_ERR, MC_BG, S);
-        printf("  XIP sentinel: %s\n", all_ok ? "PASS" : "FAIL");
-
-        // Print J-Link instructions to serial
-        printf("  J-Link command: mem32 0x%08X 4\n", (unsigned)PSRAM_NOCACHE);
-        printf("  Expected: %08X %08X %08X %08X\n",
-               (unsigned)sentinel[0], (unsigned)sentinel[1],
-               (unsigned)sentinel[2], (unsigned)sentinel[3]);
-    } else {
-        menu_str(0, 3 * CH, " XIP sentinel  SKIP", COLS, MC_HINT, MC_BG, S);
-        printf("  XIP sentinel: SKIP (init failed)\n");
-        total = 1;
-    }
-
-    // M1 register info
-    snprintf(line, sizeof(line), " M1t:%08X",
-             (unsigned)qmi_hw->m[1].timing);
-    menu_str(0, 5 * CH, line, COLS, MC_FG, MC_BG, S);
-    snprintf(line, sizeof(line), " M1r:%08X",
-             (unsigned)qmi_hw->m[1].rfmt);
-    menu_str(0, 6 * CH, line, COLS, MC_FG, MC_BG, S);
-
-    // Summary
-    snprintf(line, sizeof(line), " Result: %d/%d pass", pass, total);
-    menu_str(0, 8 * CH, line, COLS, pass == total ? MC_OK : MC_ERR, MC_BG, S);
-    menu_str(0, 10 * CH, " BACK to return     ", COLS, MC_HINT, MC_BG, S);
-
-    printf("\n  === %d/%d tests OK ===\n", pass, total);
-
-    while (!back_key_pressed()) sleep_ms(50);
-}
-
-// ---------------------------------------------------------------------------
-// PSRAM Speed Test — CPU throughput measurement
-//
-// DMA access to PSRAM through XIP is unreliable (~37.5% address errors).
-// This test uses CPU volatile access (proven correct by psram_full_test).
-//
-// Measures:
-//   1. CPU Write 8 MB (fixed pattern, timed)
-//   2. CPU Read+Verify 8 MB (timed)
-//   3. Per-word timing (ns/word, sys_clk cycles)
-// ---------------------------------------------------------------------------
-
-void cmd_psram_dma_test(void) {
-    const int S = 2;
-    const int CW = 6 * S;
-    const int CH = 8 * S;
-    const int W  = menu_tft_width();
-    const int COLS = W / CW;
-
-    menu_clear(MC_BG);
-    menu_str(0, 0, " PSRAM Speed Test   ", COLS, MC_TITLE, MC_TITBG, S);
-
-    printf("\n--- PSRAM Speed Test (CPU XIP) ---\n");
-
-    // Check PSRAM init
-    uint32_t gpio0_ctrl = *(volatile uint32_t *)0x40028004u;
-    if ((gpio0_ctrl & 0x1Fu) != 9) {
-        menu_str(0, 2 * CH, " PSRAM not init'd   ", COLS, MC_ERR, MC_BG, S);
-        menu_str(0, 3 * CH, " Run MemDiag first  ", COLS, MC_HINT, MC_BG, S);
-        menu_str(0, 10 * CH, " BACK to return     ", COLS, MC_HINT, MC_BG, S);
-        while (!back_key_pressed()) sleep_ms(50);
-        return;
-    }
-
-    volatile uint32_t *psram = (volatile uint32_t *)PSRAM_NOCACHE;
-    uint32_t pattern = 0xA5A5A5A5u;
-    char line[24];
-    uint32_t t0, t1;
-    uint32_t sys_hz = clock_get_hz(clk_sys);
-
-    // === 1. CPU Write 8 MB ===
-    menu_str(0, 2 * CH, " Writing 8 MB...    ", COLS, MC_FG, MC_BG, S);
-
-    t0 = to_ms_since_boot(get_absolute_time());
-    for (uint32_t i = 0; i < PSRAM_FULL_WORDS; i++)
-        psram[i] = pattern;
-    t1 = to_ms_since_boot(get_absolute_time());
-    uint32_t wr_ms = t1 - t0;
-    uint32_t wr_kbps = wr_ms ? (8u * 1024u * 1000u / wr_ms) : 0;
-
-    snprintf(line, sizeof(line), " Wr: %ums %uKB/s",
-             (unsigned)wr_ms, (unsigned)wr_kbps);
-    menu_str(0, 2 * CH, line, COLS, MC_FG, MC_BG, S);
-    printf("  CPU Write 8MB: %u ms  (%u KB/s)\n",
-           (unsigned)wr_ms, (unsigned)wr_kbps);
-
-    if (back_key_pressed()) goto speed_done;
-
-    // === 2. CPU Read+Verify 8 MB ===
-    menu_str(0, 3 * CH, " Verifying 8 MB...  ", COLS, MC_FG, MC_BG, S);
-
-    uint32_t errs = 0;
-    t0 = to_ms_since_boot(get_absolute_time());
-    for (uint32_t i = 0; i < PSRAM_FULL_WORDS; i++)
-        if (psram[i] != pattern) errs++;
-    t1 = to_ms_since_boot(get_absolute_time());
-    uint32_t rd_ms = t1 - t0;
-    uint32_t rd_kbps = rd_ms ? (8u * 1024u * 1000u / rd_ms) : 0;
-
-    snprintf(line, sizeof(line), " Rd: %ums %uKB/s",
-             (unsigned)rd_ms, (unsigned)rd_kbps);
-    menu_str(0, 3 * CH, line, COLS,
-             errs == 0 ? MC_FG : MC_ERR, MC_BG, S);
-    printf("  CPU Read 8MB:  %u ms  (%u KB/s)  errors=%u\n",
-           (unsigned)rd_ms, (unsigned)rd_kbps, (unsigned)errs);
-
-    // === Per-word timing ===
-    {
-        uint32_t wr_ns = wr_ms ? (uint32_t)((uint64_t)wr_ms * 1000000ULL / PSRAM_FULL_WORDS) : 0;
-        uint32_t rd_ns = rd_ms ? (uint32_t)((uint64_t)rd_ms * 1000000ULL / PSRAM_FULL_WORDS) : 0;
-        uint32_t wr_clk = wr_ns * (sys_hz / 1000000u) / 1000u;
-        uint32_t rd_clk = rd_ns * (sys_hz / 1000000u) / 1000u;
-
-        snprintf(line, sizeof(line), " Wr:%uns %uclk",
-                 (unsigned)wr_ns, (unsigned)wr_clk);
-        menu_str(0, 7 * CH, line, COLS, MC_FG, MC_BG, S);
-        snprintf(line, sizeof(line), " Rd:%uns %uclk",
-                 (unsigned)rd_ns, (unsigned)rd_clk);
-        menu_str(0, 8 * CH, line, COLS, MC_FG, MC_BG, S);
-
-        printf("  Write: %u ns/word (%u sys_clk)\n",
-               (unsigned)wr_ns, (unsigned)wr_clk);
-        printf("  Read:  %u ns/word (%u sys_clk)\n",
-               (unsigned)rd_ns, (unsigned)rd_clk);
-    }
-
-    // === Result summary ===
-    snprintf(line, sizeof(line), " %s  err=%u",
-             errs == 0 ? "PASS" : "FAIL", (unsigned)errs);
-    menu_str(0, 10 * CH, line, COLS,
-             errs == 0 ? MC_OK : MC_ERR, MC_BG, S);
-
-speed_done:
-    menu_str(0, 12 * CH, " BACK to return     ", COLS, MC_HINT, MC_BG, S);
-    while (!back_key_pressed()) sleep_ms(50);
-}
-
-// ---------------------------------------------------------------------------
-// PSRAM Read Diagnostic — investigate read vs write speed asymmetry
-//
-// Dumps QMI/XIP registers, then runs isolated read/write benchmarks
-// without verify overhead to isolate the true DMA throughput.
-// Serial-only output for debugging.
-// ---------------------------------------------------------------------------
-void psram_rd_diag(void) {
-    printf("\n=== PSRAM Read Diagnostic ===\n");
-
-    uint32_t gpio0_ctrl = *(volatile uint32_t *)0x40028004u;
-    if ((gpio0_ctrl & 0x1Fu) != 9) {
-        printf("  PSRAM not initialized. Run 'psram' first.\n");
-        return;
-    }
-
-    uint32_t sys_hz = clock_get_hz(clk_sys);
-    printf("\n  sys_clk = %u MHz\n", (unsigned)(sys_hz / 1000000u));
-
-    // --- 1. QMI M1 register dump ---
-    printf("\n--- QMI M1 Registers ---\n");
-    printf("  M1.timing  = 0x%08X\n", (unsigned)qmi_hw->m[1].timing);
-    uint32_t t = qmi_hw->m[1].timing;
-    printf("    CLKDIV=%u  RXDELAY=%u  COOLDOWN=%u  MIN_DESELECT=%u\n",
-           (unsigned)(t & 0xFF),
-           (unsigned)((t >> 8) & 0x7),
-           (unsigned)((t >> 30) & 0x3),
-           (unsigned)((t >> 22) & 0x1F));
-    printf("    SELECT_HOLD=%u  SELECT_SETUP=%u  MAX_SELECT=%u  PAGEBREAK=%u\n",
-           (unsigned)((t >> 16) & 0x3),
-           (unsigned)((t >> 18) & 0x3),
-           (unsigned)((t >> 27) & 0x7),
-           (unsigned)((t >> 20) & 0x3));
-
-    uint32_t sck_mhz = sys_hz / (2u * (t & 0xFF)) / 1000000u;
-    printf("    SCK = %u MHz\n", (unsigned)sck_mhz);
-
-    printf("  M1.rfmt    = 0x%08X\n", (unsigned)qmi_hw->m[1].rfmt);
-    uint32_t rf = qmi_hw->m[1].rfmt;
-    printf("    PREFIX_LEN=%u  PREFIX_WIDTH=%u  ADDR_WIDTH=%u\n",
-           (unsigned)(rf & 0x3),
-           (unsigned)((rf >> 2) & 0x3),
-           (unsigned)((rf >> 4) & 0x3));
-    printf("    DUMMY_LEN=%u  DUMMY_WIDTH=%u  DATA_WIDTH=%u  SUFFIX_LEN=%u\n",
-           (unsigned)((rf >> 16) & 0x7),
-           (unsigned)((rf >> 10) & 0x3),
-           (unsigned)((rf >> 8) & 0x3),
-           (unsigned)((rf >> 6) & 0x3));
-    printf("  M1.rcmd    = 0x%02X\n", (unsigned)qmi_hw->m[1].rcmd);
-
-    printf("  M1.wfmt    = 0x%08X\n", (unsigned)qmi_hw->m[1].wfmt);
-    printf("  M1.wcmd    = 0x%02X\n", (unsigned)qmi_hw->m[1].wcmd);
-
-    // --- 2. QMI M0 register dump (for comparison) ---
-    printf("\n--- QMI M0 Registers (Flash, reference) ---\n");
-    printf("  M0.timing  = 0x%08X\n", (unsigned)qmi_hw->m[0].timing);
-    uint32_t t0r = qmi_hw->m[0].timing;
-    printf("    CLKDIV=%u  RXDELAY=%u  COOLDOWN=%u  MIN_DESELECT=%u  PAGEBREAK=%u\n",
-           (unsigned)(t0r & 0xFF),
-           (unsigned)((t0r >> 8) & 0x7),
-           (unsigned)((t0r >> 30) & 0x3),
-           (unsigned)((t0r >> 22) & 0x1F),
-           (unsigned)((t0r >> 20) & 0x3));
-    printf("  M0.rfmt    = 0x%08X\n", (unsigned)qmi_hw->m[0].rfmt);
-    printf("  M0.rcmd    = 0x%02X\n", (unsigned)qmi_hw->m[0].rcmd);
-
-    // --- 3. XIP_CTRL ---
-    printf("\n--- XIP_CTRL ---\n");
-    uint32_t xc = xip_ctrl_hw->ctrl;
-    printf("  XIP_CTRL   = 0x%08X\n", (unsigned)xc);
-    printf("    EN=%u  WRITABLE_M1=%u  POWER_DOWN=%u\n",
-           (unsigned)(xc & 1),
-           (unsigned)((xc >> 10) & 1),
-           (unsigned)((xc >> 3) & 1));
-    printf("    SPLIT_WAYS=%u  MAINT_NONSEC=%u\n",
-           (unsigned)((xc >> 4) & 1),
-           (unsigned)((xc >> 5) & 1));
-
-    // Cache status
-    printf("  XIP_STAT   = 0x%08X\n", (unsigned)xip_ctrl_hw->stat);
-
-    // --- 4. ATRANS (address translation) for M1 ---
-    printf("\n--- QMI ATRANS (address translation) ---\n");
-    for (int i = 0; i < 8; i++) {
-        uint32_t at = qmi_hw->atrans[i];
-        printf("  ATRANS[%d]  = 0x%08X  (BASE=0x%06X, SIZE=%u)\n",
-               i, (unsigned)at,
-               (unsigned)((at >> 2) & 0x3FFFFF) << 12,
-               (unsigned)(at & 0x3));
-    }
-
-    // --- 5. Pure DMA write (no verify) ---
-    printf("\n--- Pure DMA Benchmarks (no verify) ---\n");
-    volatile uint32_t *psram = (volatile uint32_t *)PSRAM_NOCACHE;
-    static uint32_t pat = 0xA5A5A5A5u;
-    #define RD_DIAG_CHUNK_WORDS (64u * 1024u / 4u)  // 64 KB in words
-    static uint32_t rd_diag_buf[RD_DIAG_CHUNK_WORDS];
-
-    int dma_ch = dma_claim_unused_channel(true);
-    dma_channel_config cfg = dma_channel_get_default_config(dma_ch);
-    channel_config_set_transfer_data_size(&cfg, DMA_SIZE_32);
-
-    // DMA Write
-    channel_config_set_read_increment(&cfg, false);
-    channel_config_set_write_increment(&cfg, true);
-
-    uint32_t t0 = to_ms_since_boot(get_absolute_time());
-    dma_channel_configure(dma_ch, &cfg,
-                          (void *)psram, &pat,
-                          PSRAM_FULL_WORDS, true);
-    dma_channel_wait_for_finish_blocking(dma_ch);
-    uint32_t wr_ms = to_ms_since_boot(get_absolute_time()) - t0;
-    uint32_t wr_kbps = wr_ms ? (8u * 1024u * 1000u / wr_ms) : 0;
-    printf("  DMA Write 8MB:      %u ms  (%u KB/s)\n",
-           (unsigned)wr_ms, (unsigned)wr_kbps);
-
-    // DMA Read (pure — into SRAM, no verify, discard data)
-    channel_config_set_read_increment(&cfg, true);
-    channel_config_set_write_increment(&cfg, true);
-
-    // Read in 64 KB chunks (DMA dst = rd_diag_buf, just overwrite same buffer)
-    t0 = to_ms_since_boot(get_absolute_time());
-    for (uint32_t off = 0; off < PSRAM_FULL_WORDS; off += RD_DIAG_CHUNK_WORDS) {
-        uint32_t count = RD_DIAG_CHUNK_WORDS;
-        if (off + count > PSRAM_FULL_WORDS) count = PSRAM_FULL_WORDS - off;
-        dma_channel_configure(dma_ch, &cfg,
-                              rd_diag_buf, (void *)(psram + off),
-                              count, true);
-        dma_channel_wait_for_finish_blocking(dma_ch);
-    }
-    uint32_t rd_ms = to_ms_since_boot(get_absolute_time()) - t0;
-    uint32_t rd_kbps = rd_ms ? (8u * 1024u * 1000u / rd_ms) : 0;
-    printf("  DMA Read 8MB:       %u ms  (%u KB/s)\n",
-           (unsigned)rd_ms, (unsigned)rd_kbps);
-
-    // DMA Read single 64 KB block (no loop overhead)
-    channel_config_set_read_increment(&cfg, true);
-    channel_config_set_write_increment(&cfg, true);
-    t0 = to_ms_since_boot(get_absolute_time());
-    dma_channel_configure(dma_ch, &cfg,
-                          rd_diag_buf, (void *)psram,
-                          RD_DIAG_CHUNK_WORDS, true);
-    dma_channel_wait_for_finish_blocking(dma_ch);
-    uint32_t rd64_ms = to_ms_since_boot(get_absolute_time()) - t0;
-    uint32_t rd64_kbps = rd64_ms ? (64u * 1000u / rd64_ms) : 0;
-    printf("  DMA Read 64KB:      %u ms  (%u KB/s)\n",
-           (unsigned)rd64_ms, (unsigned)rd64_kbps);
-
-    // Per-word timing
-    printf("\n--- Per-word Timing ---\n");
-    uint32_t wr_ns = wr_ms ? (uint32_t)((uint64_t)wr_ms * 1000000ULL / PSRAM_FULL_WORDS) : 0;
-    uint32_t rd_ns = rd_ms ? (uint32_t)((uint64_t)rd_ms * 1000000ULL / PSRAM_FULL_WORDS) : 0;
-    printf("  Write: %u ns/word  (%u sys_clk @ %u MHz)\n",
-           (unsigned)wr_ns, (unsigned)(wr_ns * (sys_hz / 1000000u) / 1000u),
-           (unsigned)(sys_hz / 1000000u));
-    printf("  Read:  %u ns/word  (%u sys_clk @ %u MHz)\n",
-           (unsigned)rd_ns, (unsigned)(rd_ns * (sys_hz / 1000000u) / 1000u),
-           (unsigned)(sys_hz / 1000000u));
-    printf("  Ratio: read/write = %u.%ux\n",
-           (unsigned)(rd_ms * 10u / wr_ms / 10u),
-           (unsigned)(rd_ms * 10u / wr_ms % 10u));
-
-    // Expected QPI clocks (no burst)
-    printf("\n--- Expected (no burst, per word) ---\n");
-    printf("  Write: cmd(2)+addr(6)+data(8) = 16 QPI clk = %u ns\n",
-           (unsigned)(16u * 1000u / sck_mhz));
-    printf("  Read:  cmd(2)+addr(6)+dummy(7)+data(8) = 23 QPI clk = %u ns\n",
-           (unsigned)(23u * 1000u / sck_mhz));
-    printf("  Expected ratio: 1.44x (dummy overhead only)\n");
-    printf("  Actual ratio:   %u.%ux\n",
-           (unsigned)(rd_ms * 10u / wr_ms / 10u),
-           (unsigned)(rd_ms * 10u / wr_ms % 10u));
-
-    if (rd_ms > wr_ms * 2) {
-        printf("\n  >>> Write appears to BURST, read does NOT <<<\n");
-        printf("  Check M1.timing PAGEBREAK field:\n");
-        uint32_t pb = (t >> 20) & 0x3;
-        const char *pb_str[] = {"NONE (no break)", "256B", "512B", "1024B"};
-        printf("    PAGEBREAK = %u (%s)\n", (unsigned)pb, pb_str[pb]);
-        printf("  APS6404L page size = 1024 bytes → PAGEBREAK should be 3\n");
-    }
-
-    dma_channel_unclaim(dma_ch);
-    printf("\n=== Done ===\n");
-}
